Added Ray Casting
Adding Ray Casting to the enemy object.  As is it doesn't do anything
interesting.  My thought is to find a way to make the enemy line of
sight sweep rapidly over an arch length to simulate a vision cone.  Once
that is done, I will define behaviors for various line of sight to tag
collisions.